# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Heritage of Faith Church — audio content and subscription platform.",
        "title": "HOF Backend API",
        "termsOfService": "https://hofng.org/terms",
        "contact": {
            "name": "HOF Dev Team",
            "email": "dev@hofng.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/global": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get global application parameters",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update global application parameters",
                "parameters": [
                    {
                        "description": "Parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_subscription.UpdateGlobalParamsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/session/authenticate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh a session using refresh token",
                "parameters": [
                    {
                        "description": "Tokens",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.AuthenticateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.SessionResponse"
                        }
                    }
                }
            }
        },
...
        "/session/sign_in": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in with email and password",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.SessionResponse"
                        }
                    }
                }
            }
        },
        "/session/sign_in/admin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Admin sign in",
                "parameters": [
                    {
                        "description": "Admin credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.AdminLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.SessionResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bitbucket_org_hofng_hofApp_internal_application_auth.AdminLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "bitbucket_org_hofng_hofApp_internal_application_auth.AuthenticateRequest": {
            "type": "object",
            "required": [
                "refresh_token",
                "token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "bitbucket_org_hofng_hofApp_internal_application_auth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "device_identifier": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "bitbucket_org_hofng_hofApp_internal_application_auth.SessionResponse": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "subscription": {
                    "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.SubscriptionDTO"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/bitbucket_org_hofng_hofApp_internal_application_auth.UserDTO"
                }
            }
        },
        "bitbucket_org_hofng_hofApp_internal_application_auth.SubscriptionDTO": {
            "type": "object",
            "properties": {
                "next_payment_date": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "bitbucket_org_hofng_hofApp_internal_application_auth.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "bitbucket_org_hofng_hofApp_internal_application_subscription.UpdateGlobalParamsRequest": {
            "type": "object",
            "properties": {
                "activate_subscription": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}